cnc-mill-app/python-application/requirements.txt
wheel==0.37.0
anchor-exp==0.0.2.0
Flask==1.1.4
prometheus-client==0.11.0
Flask-MonitoringDashboard==3.1.0
flask-prometheus-metrics==1.0.0
GitPython==3.1.18
ipython==7.27.0
Jinja2==2.11.1
joblib==1.0.1
lime==0.2.0.1
matplotlib==3.4.2
mlflow==1.20.1
numpy==1.20.3
pandas==1.3.1
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.1
shap==0.39.0
tzlocal==2.1
Werkzeug==1.0.1
auto-sklearn==0.13.0